PS2 Port working, will post external libs soon, hopefully source code for them will follow, if I find that Sony's OpenGL library is horrid.
There are also changes to software renderer and iG3D interface regarding
uint32 ULong issues.  You should not assume that ULong is a 32 bit datatype,
but rather a datatype thats big and goes really fast on the target platform.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@6504 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/cstypes.h
@@ -65,8 +65,8 @@ typedef unsigned char uint8;
 typedef char int8;
 typedef unsigned short uint16;
 typedef short int16;
-typedef unsigned long uint32;
-typedef long int32;
+typedef unsigned int uint32;
+typedef int int32;
 #endif
 
 //------------------------------
@@ -78,7 +78,7 @@ typedef unsigned int uint;
 typedef unsigned long ulong;
 
 // Time in milli (1/1000) seconds
-typedef unsigned long cs_time;
+typedef unsigned int cs_time;
 
 // A pointer to something
 typedef void *csSome;
