+++ plugins/video/renderer/glide/g3dgl.h
@@ -237,7 +237,7 @@ public:
   virtual long GetRenderState (G3D_RENDERSTATEOPTION);
 
   /// Get a z-buffer point
-  virtual unsigned long *GetZBuffAt (int, int)
+  virtual uint32 *GetZBuffAt (int, int)
   { return NULL; }
 
   /// Get the width
