+++ plugins/video/renderer/line/line_g3d.h
@@ -163,7 +163,7 @@ public:
   { return &Caps; }
 
   /// Get address of Z-buffer at specific point.
-  virtual unsigned long *GetZBuffAt (int, int) { return NULL; }
+  virtual uint32 *GetZBuffAt (int, int) { return NULL; }
 
   /// Dump the texture cache.
   virtual void DumpCache () { }
